+++ tests/test_end_to_end.py
@@ -83,13 +83,8 @@ def test_automatic_tasks(
 
     # limit number of subjects to three
     task = get_task(task_name)
-    subjects_subset = task.SUBJECTS[:3]
-
-    with patch(f"{task.__module__}.{task.__name__}.SUBJECTS", new=subjects_subset):
-        results = main(test_llms, eval_config)
-
-    # un-comment for debugging
-    # pretty_print_results(results)
+    eval_config.task_subjects = task.SUBJECTS[:3]
+    results = main(test_llms, eval_config)
 
     full_metric_names = [
         (result.metric_name, result.key) for result in results if result.metric_name in list(expected_results.keys())